requirements.txt
asgiref==3.3.4
Django==3.2
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.1
PyYAML==6.0
sqlparse==0.4.1
whitenoise==5.2.0
